refactor: Update widget implementations for improved readability and consistency
- Refactored various widget methods to enhance code clarity by adjusting parameter formatting and spacing.
- Updated accessibility-related methods to utilize new utility functions for better maintainability.
- Cleaned up unused imports and ensured consistent use of hooks across components.
- Enhanced error handling and state management in text fields and buttons for improved user experience.

+++ lib/core/responsive/responsive_utils.dart
@@ -68,7 +68,7 @@ class ResponsiveUtils {
     T? largeDesktop,
   }) {
     final breakpoints = ResponsiveBreakpoints.of(context);
-    
+
     if (breakpoints.largerThan(DESKTOP) && largeDesktop != null) {
       return largeDesktop;
     } else if (breakpoints.largerThan(TABLET) && desktop != null) {
@@ -165,7 +165,7 @@ class ResponsiveUtils {
       tablet: 12,
       desktop: 16,
     );
-    return BorderRadius.circular(radius);
+    return BorderRadius.circular(radius as double);
   }
 
   /// Get responsive elevation
